4.17第六次提交 检索功能完成了一部分: 根据条件查询,图片,高亮完成 前端页面分页有问题

+++ taotao-manager/taotao-manager-service/src/main/java/com/itszt/taotao/manager/service/GoodsServiceImpl.java
@@ -43,7 +43,7 @@ public class GoodsServiceImpl implements GoodsService {
     @Override
     @Transactional
     public boolean addGoods(TbItem tbItem, String desc) {
-        tbItem.setStatus((byte)0);
+        tbItem.setStatus((int)0);
         tbItem.setCreated(new Date());
         tbItem.setUpdated(new Date());
         boolean b = goodsDao.insertTbItem(tbItem);
